env-file: made a start with using the new docker-compose functionality
Made a start with using the new docker-compose functionality of using
and env_file key in the docker-compose.yml file. This will allow you
to specify multiple environment variables in a seperate file. This is
great if you have a lot of them, or if you want to keep sensitive
information out of version control.