config: added database check to config
Added script database-check.py to the startup procedure. docker-compose
will asynchronously spin up the containers. However the app container
will execute some commands in it's startup that needs the running
postgres container. Otherwise it will exit with an error stating it
can't connect to the database. This script will check whether a
connection can be made and will continue setting up when this succeeds.